% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapfac.R
\name{calcCapfac}
\alias{calcCapfac}
\title{Calculate capacity factors}
\usage{
calcCapfac(pyDir, py2rmTech, py2rmRegi, iter)
}
\arguments{
\item{pyDir}{PyPSA output folder with sub-folders for years}

\item{py2rmTech}{PyPSA to REMIND technology mapping}

\item{py2rmRegi}{PyPSA to REMIND region mapping}

\item{iter}{Current iteration}
}
\value{
Tibble with capacity factors for various years.
}
\description{
Calculate capacity factors
}
