% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCosts.R
\name{calcCosts}
\alias{calcCosts}
\title{Calculate capital and marginal costs (REMIND to PyPSA)
This function writes costs_yXXXX.csv files for all years.}
\usage{
calcCosts(rmFile, outDir, years, rm2pyTech, py2aggTech)
}
\arguments{
\item{rmFile}{REMIND output gdx file}

\item{outDir}{Output folder for costs_yXXXX.csv files}

\item{years}{Years coupled from REMIND to PyPSA}

\item{rm2pyTech}{REMIND to aggregated PyPSA technology mapping}

\item{py2aggTech}{PyPSA to aggregated PyPSA technology mapping}
}
\value{
Tibble with capital and marginal costs
}
\description{
Calculate capital and marginal costs (REMIND to PyPSA)
This function writes costs_yXXXX.csv files for all years.
}
