% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPowerplants.R
\name{calcPowerplants}
\alias{calcPowerplants}
\title{Calculate database of powerplants for different years (REMIND to PyPSA)
This function scales the PyPSA powerplant database so that the sum
matches the pre-investment capacities of REMIND.}
\usage{
calcPowerplants(rmFile, pyPowerplants, rm2pyTech, outDir, years)
}
\arguments{
\item{rmFile}{REMIND output gdx file}

\item{pyPowerplants}{PyPSA powerplants.csv file}

\item{rm2pyTech}{REMIND to aggregated PyPSA technology mapping}

\item{outDir}{Output folder for powerplants_yXXXX.csv files}

\item{years}{Years coupled from REMIND to PyPSA}
}
\value{
Writes one csv file for each year into outDir.
}
\description{
Calculate database of powerplants for different years (REMIND to PyPSA)
This function scales the PyPSA powerplant database so that the sum
matches the pre-investment capacities of REMIND.
}
