% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPyPSA2REMIND.R
\name{callPyPSA2REMIND}
\alias{callPyPSA2REMIND}
\title{Wrapper function to exchange data from PyPSA to REMIND}
\usage{
callPyPSA2REMIND(pyDir, iter)
}
\arguments{
\item{pyDir}{PyPSA folder}

\item{iter}{Current iteration}
}
\value{
Writes data automatically.
}
\description{
Wrapper function to exchange data from PyPSA to REMIND
}
