% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callREMIND2PyPSA.R
\name{callREMIND2PyPSA}
\alias{callREMIND2PyPSA}
\title{Wrapper function to exchange data from REMIND to PyPSA}
\usage{
callREMIND2PyPSA(pyDir, iter)
}
\arguments{
\item{pyDir}{PyPSA folder}

\item{iter}{Current iteration}
}
\value{
Writes data automatically.
}
\description{
Wrapper function to exchange data from REMIND to PyPSA
}
