% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createValidation.R
\name{createValidation}
\alias{createValidation}
\title{Create REMIND-PyPSA validation file
This function renders the /inst/rmd/REMIND-PyPSA_Validation.Rmd file.
It has to be called from the REMIND output directory.}
\usage{
createValidation()
}
\value{
REMIND-PyPSA validation file (usually in PDF format)
}
\description{
Create REMIND-PyPSA validation file
This function renders the /inst/rmd/REMIND-PyPSA_Validation.Rmd file.
It has to be called from the REMIND output directory.
}
