% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remindPypsa-package.R
\docType{package}
\name{remindPypsa-package}
\alias{remindPypsa}
\alias{remindPypsa-package}
\title{remindPypsa: REMIND-PyPSA model coupling scripts}
\description{
This package contains functions for the iterative bi-directional model coupling of the integrated assessment model REMIND and the spatio-temporally detailed power and energy system model family PyPSA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/remindPypsa}
}

}
\author{
\strong{Maintainer}: Adrian Odenweller \email{adrian.odenweller@pik-potsdam.de}

}
