% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startPyPSA.R
\name{startPyPSA}
\alias{startPyPSA}
\title{Start PyPSA}
\usage{
startPyPSA(pyDir, iter, copyConfig = TRUE)
}
\arguments{
\item{pyDir}{PyPSA folder}

\item{iter}{Current iteration}

\item{copyConfig}{Logical indicating whether to copy PyPSA config files from inst/extdata}
}
\value{
None.
}
\description{
Start PyPSA
}
