% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCapfac.R
\name{plotCapfac}
\alias{plotCapfac}
\title{Plot capacity factors}
\usage{
plotCapfac(resDir, years, rm2pyTech, colorsTech, namesTech)
}
\arguments{
\item{resDir}{REMIND directory}

\item{years}{Years coupled for REMIND-PyPSA}

\item{rm2pyTech}{REMIND to aggregated PyPSA technology mapping}

\item{colorsTech}{Colour mapping}

\item{namesTech}{Pretty names mapping}
}
\value{
List of ggplot2 objects
}
\description{
Plot capacity factors
}
