% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCosts.R
\name{plotCosts}
\alias{plotCosts}
\title{Plot capital and marginal costs from remind2pypsa.rds}
\usage{
plotCosts(remind2pypsa, colorsTech, namesTech)
}
\arguments{
\item{remind2pypsa}{rds data file}

\item{colorsTech}{Colour mapping}

\item{namesTech}{Pretty names mapping}
}
\value{
List of ggplot2 objects
}
\description{
Plot capital and marginal costs from remind2pypsa.rds
}
