% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGDXtibble.R
\name{readGDXtibble}
\alias{readGDXtibble}
\title{Read data from gdx file and return as tibble with filtered columns.}
\usage{
readGDXtibble(
  rmFile,
  gdxVar,
  columns,
  colFilter,
  restoreZeros = TRUE,
  field = "l",
  recalcUnit = 1
)
}
\arguments{
\item{rmFile}{Path to gdx file.}

\item{gdxVar}{Variable name in gdx file.}

\item{columns}{Named vector of columns in gdx file to new column names.}

\item{colFilter}{Named list of new column names to filter values of that column.}

\item{restoreZeros}{Logical indicating whether to restore zeros in readGDX.}

\item{field}{Field of variable to read from gdx file.}

\item{recalcUnit}{Factor that values are multiplied with.}
}
\value{
Tibble with renamed and filtered columns.
}
\description{
Read data from gdx file and return as tibble with filtered columns.
}
